replacing async tasks with executors
