+++ app/build.gradle
@@ -34,4 +34,5 @@ dependencies {
     implementation 'com.squareup.picasso:picasso:2.71828'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'androidx.activity:activity:1.8.2'
+    implementation 'androidx.preference:preference:1.2.1'
 }
