+++ app/src/main/java/com/droidheat/amoledbackgrounds/DailyJobService.java
@@ -6,33 +6,36 @@ import android.content.Intent;
 import android.os.Build;
 import android.os.StrictMode;
 
+import com.droidheat.amoledbackgrounds.utils.AppUtils;
+import com.droidheat.amoledbackgrounds.utils.SharedPrefsUtils;
+
 public class DailyJobService extends JobService {
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-        StrictMode.ThreadPolicy policy = new StrictMode.ThreadPolicy.Builder().permitAll().build();
-        StrictMode.setThreadPolicy(policy);
-    }
-
-    @Override
-    public boolean onStartJob(JobParameters params) {
-        AppUtils.scheduleJob(this);
-
-        if ((new SharedPrefsUtils(this))
-                .readSharedPrefsBoolean("daily_wallpaper", false)) {
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-                startForegroundService(new Intent(this, DailyWallpaperService.class));
-            } else {
-                startService(new Intent(this, DailyWallpaperService.class));
-            }
-        }
-        return false;
-    }
-
-    @Override
-    public boolean onStopJob(JobParameters params) {
-        return false;
-    }
-
+	
+	@Override
+	public void onCreate() {
+		super.onCreate();
+		StrictMode.ThreadPolicy policy = new StrictMode.ThreadPolicy.Builder().permitAll().build();
+		StrictMode.setThreadPolicy(policy);
+	}
+	
+	@Override
+	public boolean onStartJob(JobParameters params) {
+		AppUtils.scheduleJob(this);
+		
+		if ((new SharedPrefsUtils(this))
+						.readSharedPrefsBoolean("daily_wallpaper", false)) {
+			if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+				startForegroundService(new Intent(this, DailyWallpaperService.class));
+			} else {
+				startService(new Intent(this, DailyWallpaperService.class));
+			}
+		}
+		return false;
+	}
+	
+	@Override
+	public boolean onStopJob(JobParameters params) {
+		return false;
+	}
+	
 }
