--- app/src/main/java/com/droidheat/amoledbackgrounds/GenericFileProvider.java
@@ -1,6 +1,0 @@
-package com.droidheat.amoledbackgrounds;
-
-import androidx.core.content.FileProvider;
-
-public class GenericFileProvider extends FileProvider {
-}
